+++ Runtime/CornerHeaderCell.cs
@@ -1,0 +1,9 @@
+﻿namespace Nonatomic.UIElements
+{
+	public class CornerHeaderCell : HeaderCell
+	{
+		public CornerHeaderCell(string text, float width, float height) : base(text, width, height)
+		{
+		}
+	}
+}