+++ Runtime/HeaderCell.cs
@@ -17,6 +17,11 @@ namespace Nonatomic.UIElements
 			_label = new Label(text);
 			_label.AddToClassList("ui-table__cell-label");
 			Add(_label);
+			
+			RegisterCallback<PointerEnterEvent>(evt => OnCellPointerEnter());
+			RegisterCallback<PointerLeaveEvent>(evt => OnCellPointerLeave());
+			RegisterCallback<PointerDownEvent>(evt => OnCellPointerDown());
+			RegisterCallback<PointerUpEvent>(evt => OnCellPointerUp());
 		}
 
 		public void SetLabel(string text)
@@ -27,6 +32,29 @@ namespace Nonatomic.UIElements
 		public void SetWidth(float width)
 		{
 			style.width = width;
+		}
+		
+		protected void OnCellPointerUp()
+		{
+			//AddToClassList("ui-table__cell--highlighted");
+			//RemoveFromClassList("ui-table__cell--down");
+		}
+
+		protected void OnCellPointerDown()
+		{
+			//RemoveFromClassList("ui-table__cell--highlighted");
+			//AddToClassList("ui-table__cell--down");
+		}
+		
+		protected void OnCellPointerEnter()
+		{
+			//AddToClassList("ui-table__cell--highlighted");
+		}
+
+		protected void OnCellPointerLeave()
+		{
+			//RemoveFromClassList("ui-table__cell--highlighted");
+			//RemoveFromClassList("ui-table__cell--down");
 		}
 	}
 }